+++ run_tests.py
@@ -1,15 +1,17 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 if __name__ == '__main__':
     import pytest
     import sys
     package_name = 'torch_liberator'
+    mod_dpath = package_name
+    test_dpath = 'tests'
     pytest_args = [
-        '--cov-config', '.coveragerc',
+        '--cov-config', 'pyproject.toml',
         '--cov-report', 'html',
         '--cov-report', 'term',
         '--xdoctest',
         '--cov=' + package_name,
+        mod_dpath, test_dpath
     ]
     pytest_args = pytest_args + sys.argv[1:]
     sys.exit(pytest.main(pytest_args))