+++ torch_liberator/__init__.py
@@ -1,15 +1,15 @@
 """
-mkinit ~/code/torch_liberator/torch_liberator/__init__.py -w
+mkinit torch_liberator -w
 """
 
-__version__ = '0.1.1'
+__version__ = '0.2.0'
 
 from torch_liberator import deployer
 from torch_liberator import exporter
 
 from torch_liberator.deployer import (DeployedModel, deploy,)
 from torch_liberator.exporter import (export_model_code,)
-from torch_liberator.initializer import (load_partial_state,)
+from torch_liberator.initializer import (load_partial_state, Pretrained)
 
 __all__ = ['DeployedModel', 'deploy', 'deployer', 'export_model_code',
-           'exporter', 'load_partial_state']
+           'exporter', 'load_partial_state', 'Pretrained']